requirements.txt
albumentations==2.0.5
loguru==0.7.3
matplotlib==3.10.0 
protobuf==4.25.3 
pyarrow==19.0.1
python-dotenv==1.0.1 
PyYAML==6.0.2
ray==2.43.0 
setuptools==75.1.0
torch==2.5.1 
tqdm==4.67.1 
typer==0.15.1
ultralytics==8.3.84 
wandb==0.19.7 